Adicionando controle de usuario e editando tela e edicao

+++ src/app/components/home/home.component.ts
@@ -5,6 +5,8 @@ import { LoginService } from '../../services/login/login.service';
 import { DocumentationService } from '../../services/api/documentation.service';
 import { CookieService } from 'ngx-cookie-service';
 import { ImageService } from '../../services/image.service';
+import { Documentation } from '../documentation/documentation';
+import { User } from '../../models/user';
 
 @Component({
   selector: 'app-home',
@@ -12,6 +14,10 @@ import { ImageService } from '../../services/image.service';
   styleUrl: './home.component.css'
 })
 export class HomeComponent implements OnInit {
+
+
+  documentationUserLists: any = []
+  documentationEdit:Documentation = new Documentation('',new User('','',''))
 
 
   listDocumentations : any = []
@@ -27,7 +33,63 @@ export class HomeComponent implements OnInit {
     , private imageService : ImageService
   ){}
 
-  carregarDocumentacao(documentation:any){
+
+  addUserDocumentation(idUser:any){
+    const retorno = this.documentationService.addUsersDocumentation(this.documentationEdit,idUser)
+    retorno.subscribe(
+      response=>{
+        console.log(response);
+        
+      }
+      ,error=>{
+        console.log(error);
+        
+      }
+    )
+  }
+
+  carregarListaUsuarios(documentation:Documentation){
+
+    this.documentationEdit = documentation
+
+    const retorno = this.documentationService.selectUsersDocumentation(documentation);
+    retorno.subscribe(
+      response=>{
+        let result: User[] = []
+        response.forEach((element: any) => {
+          let user = new User(
+              ''
+              ,''
+              ,element.nome
+            )
+
+            user.funcao = element.funcao
+            user.id = element.id
+            
+            result.push(user)
+        });
+        console.log(result);
+        
+        this.documentationUserLists = result
+
+      },
+      error=>{
+        console.log(error.error);
+        
+      }
+    )
+
+  }
+
+
+
+
+
+
+
+
+
+  carregarDocumentacao(documentation:Documentation){
     this.storageAcessService.updateStorage(documentation)
     this.router.navigate(['/documentation/edit'])
   }
@@ -36,17 +98,20 @@ export class HomeComponent implements OnInit {
   deleteDocumentation(documentation:any){
     console.log(documentation);
 
-    const retorno = this.documentationService.deleteDocumentation(documentation)
-    retorno.subscribe(
-      result=>{
-        console.log(result);
-        this.ngOnInit()
-      }
-      ,error=>{
-        console.log(error);
-        
-      }
-    )
+
+    if (confirm('Deseja mesmo deletar a documentação : '+documentation.titulo)) {
+      const retorno = this.documentationService.deleteDocumentation(documentation)
+      retorno.subscribe(
+        result=>{
+          console.log(result);
+          this.ngOnInit()
+        }
+        ,error=>{
+          console.log(error);
+          
+        }
+      )
+    }
     
   }
 
@@ -118,7 +183,22 @@ export class HomeComponent implements OnInit {
         console.log(response);
         let docs:any =[]
         response.forEach((element: any) => {
-          docs.push(element)
+          let doc = new Documentation(
+            element.titulo
+            , element.usuarioAlteracao
+            , element.descricao
+            , element.abas
+            , element.commitText
+            , []
+            , element.versao
+            , element.status
+            , element.dataAlteracao
+          )
+          doc.id = element.id
+          doc.imagemCapa = element.imagemCapa
+          doc.tags = element.tags
+          docs.push(doc)
+          
         });
 
         console.log(docs);
